+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Rikulo Commons Changes
 
+### 4.6.0
+
+* BREAK CHANGE: `configureDefers()`'s `executor` argument's signature is changed.
+* BREAK CHANGE: `defer`'s `categoryKey` is renamed to `category`
+
 ### 4.5.0
 
 * `configureDefers()` supports `executable` to slow down if the system is busy. 
 * BREAK CHANGE: `name` is removed from `Browser`
-* BREAK CHANGE: `onStatusCode` is remvoed from `ajax`
+* BREAK CHANGE: `onStatusCode` is removed from `ajax`
 
 ### 4.3.3